+++ packages/tsurlfilter/CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version to the end of the file -->
 <!-- e.g. [1.0.77]: https://github.com/AdguardTeam/tsurlfilter/compare/tsurlfilter-v1.0.76...tsurlfilter-v1.0.77 -->
 
+## [2.1.2] - 2023-06-19
+
+### Fixed
+
+- unwanted escape character removal when running `RuleConverter.convertRule()` on rules with regexp-value modifiers
+
 ## [2.1.1] - 2023-06-16
 
 ### Added